+++ scripts/xmlDescription.py
@@ -1,7 +1,7 @@
 # Genterated automatically by /run.ipynb
 from pydantic_xml import BaseXmlModel, attr, element
 from pydantic import Field
-from typing import Optional, Union, List
+from typing import Optional, Union, List, Literal
 
 class IDModel(BaseXmlModel):
     id: str = attr()
@@ -11,8 +11,25 @@ class IO(IDModel):
 class Ports(BaseXmlModel, tag="ports"):
     input: List[IO] = element()
     output: List[IO] = element()
+
+class Parameter(IDModel):
+    description: Optional[str] = None
+class ParameterDescription(BaseXmlModel, tag="parameter_description"):
+    parameter: List[Parameter] = element()
+
+class PartialLogicDescription(IDModel):
+    description: Optional[str] = None
+    width_description: Optional[str] = attr(default=None)
+    type: Literal["combinational_logic", "combinational_logic_operation", "sequential_logic", "sequential_logic_operation"] = attr()
+
+class LogicDescription(BaseXmlModel, tag="logic_description", search_mode='unordered'):
+    description: str = element()
+    logic: List[PartialLogicDescription] = element(default=None)
+
 class Module(IDModel, tag='module'):
     description: str = element(default=None)
     ports: Optional[Ports] = element(default=None)
+    parameter_description: Optional[ParameterDescription] = element(default=None)
+    logic_description: Optional[LogicDescription] = element(default=None)
     implementation: str = element()
     task: str = element(default="Give me the complete Verilog code.")